requirements.txt
alembic==1.11.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Classful==0.14.2
Flask-JSGlue2==0.3.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Mobility==1.1.0
Flask-Moment==1.0.5
Flask-Obscure==1.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.1
Mako==1.2.4
MarkupSafe==2.1.3
numpy==2.0.0
obscure==1.0.1
packaging==23.1
Pillow==10.0.0
potrace-cli==0.0.4
potracer==0.0.4
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-dotenv==1.0.0
qrcode==7.4.2
shortuuid==1.0.11
SQLAlchemy==2.0.18
typing_extensions==4.7.1
Werkzeug==2.3.6
wfastcgi==3.0.0
WTForms==3.0.1
